Lift Movies & SearchBar states to VideoStore
