+++ src/components/SearchBar.js
@@ -4,13 +4,6 @@ import PropTypes from 'prop-types';
 import './SearchBar.css';
 
 class SearchBar extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      queryString: ""
-    };
-  }
 
   queryChanged = (event) => {
     const queryString = event.target.value;
@@ -26,7 +19,7 @@ class SearchBar extends Component {
         <label>
           Seach <input name="search"
                         type="text"
-                        value={this.state.queryString}
+                        value={this.props.queryString}
                         onChange={this.queryChanged}
                         className="search-bar"
                         />
