+++ src/components/VideoStore.js
@@ -1,9 +1,55 @@
 import React, { Component } from 'react';
+import axios from 'axios';
 
 import Movies from './Movies';
 import SearchBar from './SearchBar';
 
 class VideoStore extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      movieList: [],  
+      currentMovie: undefined,
+      queryString: ""
+    };
+  }
+
+  url = "https://enigmatic-chamber-40825.herokuapp.com/"
+
+  componentDidMount() {
+    axios.get(this.url)
+    .then((response) => {
+      console.log(response.data);
+
+      const movieList = response.data.map((movie) => {
+        const newMovie = {
+          id: movie.id,
+          title: movie.title,
+          overview: movie.overview,
+          release_date: movie.release_date,
+          image_url: movie.image_url,
+          external_id: movie.external_id,
+        }
+        return newMovie;
+      })
+
+      console.log(movieList);
+
+      this.setState({ movieList });
+    })
+    .catch((error) => {
+      this.setState({ error: error.message })
+    })
+  }
+
+  onMovieSelect = (movieID) => {
+    const currentMovie = this.state.movieList.filter(movie => movie.id === movieID)[0]
+
+    console.log(currentMovie)
+    this.setState({ currentMovie });
+  };
+
   searchCallback = (queryString) => {
     this.setState({ queryString });
   }
@@ -11,8 +57,8 @@ class VideoStore extends Component {
   render() {
     return (
       <div>
-        <SearchBar searchCallback={this.searchCallback} />
-        <Movies queryString={this.props.queryString} />
+        <SearchBar searchCallback={this.searchCallback} onChange={this.queryChanged} queryString={this.state.queryString} />
+        <Movies movieList={this.state.movieList} currentMovie={this.state.currentMovie} onMovieSelect={this.onMovieSelect} queryString={this.state.queryString} />
       </div>
     )
   }
